refactor: move ucan service methods to own package
